CSS3 2D Transforms: done
git-svn-id: http://sources.disruptive-innovations.com/bluegriffon/trunk@215 c25cacc5-b657-0410-8d04-e2f7359d2943

+++ sidebars/cssproperties/content/cssproperties.js
@@ -69,6 +69,9 @@ function Inspect()
 
 function SelectionChanged(aArgs, aElt, aOneElementSelected)
 {
+  if (aElt == gCurrentElement)
+    return;
+
   gCurrentElement = aElt;
 
   deleteAllChildren(gDialog.classPickerPopup);
